requirements.txt
streamlit
pandas
yfinance
backtrader
datetime
matplotlib
prophet
python-dateutil
requests